Compliance check improvements and bug fixes

+++ examples/ttn-otaa-feather-us915-dht22/ttn-otaa-feather-us915-dht22.ino
@@ -112,14 +112,14 @@ void onEvent (ev_t ev) {
               Serial.print("devaddr: ");
               Serial.println(devaddr, HEX);
               Serial.print("artKey: ");
-              for (int i=0; i<sizeof(artKey); ++i) {
+              for (size_t i=0; i<sizeof(artKey); ++i) {
                 if (i != 0)
                   Serial.print("-");
                 Serial.print(artKey[i], HEX);
               }
               Serial.println("");
               Serial.print("nwkKey: ");
-              for (int i=0; i<sizeof(nwkKey); ++i) {
+              for (size_t i=0; i<sizeof(nwkKey); ++i) {
                       if (i != 0)
                               Serial.print("-");
                       Serial.print(nwkKey[i], HEX);
